groq-builder: feature(compat): extracted primitives to separate file
