+++ packages/groq-builder/src/validation/index.ts
@@ -1,4 +1,4 @@
-import { validate } from "./validate";
+import { primitives } from "./primitives";
 import {
   createGroqBuilder as _createGroqBuilder,
   RootConfig,
@@ -6,10 +6,13 @@ import {
 } from "../index";
 
 export * from "../index";
-export { validate } from "./validate";
 
 export function createGroqBuilderWithValidation<TRootConfig extends RootConfig>(
   options?: GroqBuilderOptions
 ) {
   return Object.assign(_createGroqBuilder<TRootConfig>(options), validate);
 }
+
+export const validate = {
+  ...primitives,
+};
